requirements.txt
## === Data Extract ===
pdfplumber
spacy
pandas
openpyxl

## === Agent ===
langchain
langchain-openai

## === ENV ===
python-dotenv

-e .